Q-learning
Starting to do Q-learning, the little elf moves up it keeps moving in the same direction and staying at a certain corner. Next class I might try to add something only allows the little elf to move directions that is only possible on the board so it doesn't get stuck and runs more quickly.