SARSA
Today I tried doing the SARSA dataset, from my testing there was a lot random movement even though I change my epsilon values multiple, it was able to hit the present multiple times in a row but not as consitently as q-learning. The current plan is still using q-learning for pitch day.